+++ src/ControlPanel/ControlPanel.BusinessLogic.Implementations.Tests/Menu/MenuBusinessLogic_DeleteAsync_Should.cs
@@ -1,18 +1,34 @@
-namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests.Tests.Menu
+namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests.Menu
 {
+    using System;
     using System.Collections.Generic;
     using System.Reflection;
     using System.Threading.Tasks;
+
+    using Exceptions.Menu;
 
     using FluentAssertions;
 
     using Xunit;
 
     [Collection(nameof(AssemblyName.GetAssemblyName))]
-    public class MenuBusinessLogic_GetAllAsync_Should : TestBase
+    public class MenuBusinessLogic_DeleteAsync_Should : TestBase
     {
         [Trait("Category", "Unit")]
-        public async Task Return_AllItems()
+        public async Task Throw_DeleteAsyncOperationException_WhenInputIsNull()
+        {
+            // Arrange
+
+            // Act
+            Func<Task> action = async () => { await MenuBusinessLogic.DeleteAsync(null).ConfigureAwait(false); };
+
+            // Assert
+            action.Should().ThrowExactly<MenuBusinessLogicDeleteAsyncOperationException>()
+               .WithInnerException<MenuNullInputException>();
+        }
+
+        [Trait("Category", "Unit")]
+        public async Task Delete_AnItem()
         {
             // Arrange
             DomainModel.Entities.Module module = new DomainModel.Entities.Module
@@ -44,10 +60,11 @@ namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests.Tests.
                 await MenuBusinessLogic.AddAsync(menuInactive).ConfigureAwait(false);
 
             // Act
+            await MenuBusinessLogic.DeleteAsync(menuInactiveResult).ConfigureAwait(false);
             List<DomainModel.Entities.Menu> result = await MenuBusinessLogic.GetAllAsync().ConfigureAwait(false);
 
             // Assert
-            result.Count.Should().Be(2);
+            result.Count.Should().Be(1);
         }
     }
 }
