+++ src/ControlPanel/ControlPanel.BusinessLogic.Implementations.Tests/TestBase.cs
@@ -1,4 +1,4 @@
-namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests.Tests
+namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests
 {
     using Ctx.Ctx;
